Rewrite localization system
Localizations are now managed by a few functions in the addon table. The
current locale can be retrieved with addonTable.GetLocale(), which
returns an auto-updating locale table. New locales can be created with
addonTable.NewLocale(). Localizations should be populated using the
lua_additive_table format, so as to not break unpackaged sources.

This also gets rid of the global L and LOCALE names, which were
dangerous (due to potential addon conflicts).

Added a default_enUS.lua localization that is a hard copy of the current
English localizations, and is only included in debug builds. This
provides an english localization for unpackaged sources.